Added input [macosx]

+++ engine/include/platform/system.h
@@ -5,6 +5,6 @@
 
 int init_system();
 uint64_t get_current_time();
-uint8_t system_key_to_rge_key(uint8_t system_key);
+uint8_t system_key_to_rge_key(uint16_t system_key);
 
 #endif
